requirements.txt
adafruit-circuitpython-motorkit
pygame
opencv-python
face-recognition
ttkbootstrap
Pillow
pyserial